- Added BattlePresenterInterface, implemented by BattlePresenter, and BattleInteractor use case class in order for Battle code structure to follow clean architecture. - Added code to handle the case where player enters combat with one or more of their characters fainted. - endBattle() has a partial implementation in BattlePresenter, unfinished. - endBattle(), _addReward(), and _removeActiveBattle() are unimplemented in BattleInteractor
