+++ src/main/java/com/mg105/interface_adapters/BattlePresenter.java
@@ -1,29 +1,23 @@
 package com.mg105.interface_adapters;
 
-import com.mg105.entities.Battle;
-import com.mg105.entities.BattleCharacter;
-import com.mg105.entities.GameState;
-import com.mg105.entities.Move;
-import com.mg105.user_interface.BattleMenu;
+import com.mg105.presenter_interfaces.BattlePresenterInterface;
+import com.mg105.use_cases.BattleInteractor;
 
 import java.util.ArrayList;
-import java.util.Random;
 
 /**
  * Class communicating between BattleMenu view and game entities.
  * Follows MVP pattern; contains a reference to the view and interacts with Battle and related classes.
  */
-public class BattlePresenter {
+public class BattlePresenter implements BattlePresenterInterface {
 
-    private final GameState state;
-    private Battle encounter;
+    private final BattleInteractor interactor;
 
     private BattleMenuInterface view;
 
-    public BattlePresenter(GameState state) {
-        this.state = state;
-        this.encounter = state.getCurrEncounter();
-        BattleMenu.setPresenter(this);
+    public BattlePresenter(BattleInteractor interactor) {
+        this.interactor = interactor;
+        this.interactor.setPresenter(this);
     }
 
     /**
@@ -34,36 +28,21 @@ public class BattlePresenter {
     public void setView(BattleMenuInterface view) {
         this.view = view;
     }
+
     /**
      * Creates a new encounter with random opponents and sets it as the current encounter in GameState
      */
     public void createEncounter() {
-        Random rand = new Random();
+        interactor.createEncounter();
+    }
 
-        ArrayList<BattleCharacter> opponents = new ArrayList<>();
-
-        for (int i = 0; i < 4; ++i){
-            int charHealth = rand.nextInt(5, 41);
-            int charDmg = rand.nextInt(1, 11);
-            int charSpeed = rand.nextInt(3, 16);
-            Move m1 = new Move(-rand.nextInt(1, 8), 0, "first", false);
-            Move m2 = new Move(rand.nextInt(1, 4), 0, "second", true);
-            BattleCharacter character = new BattleCharacter(charHealth, "Opponent " + i,
-                charDmg, charSpeed, true, m1, m2);
-            opponents.add(character);
-        }
-        ArrayList<BattleCharacter> party = this.state.getParty();
-        Battle b = new Battle(opponents, party);
-        this.state.setCurrEncounter(b);
-        this.encounter = b;
-
-        String[] partyNames = new String[party.size()];
-        String[] opponentNames = new String[opponents.size()];
-
-        for (int i = 0; i < party.size(); ++i) {
-            partyNames[i] = party.get(i).getName();
-            opponentNames[i] = opponents.get(i).getName();
-        }
+    /**
+     * Set the characters in the view.
+     * @param partyNames the name strings of the party characters.
+     * @param opponentNames the name strings of the opponent characters.
+     */
+    @Override
+    public void setViewNames(String[] partyNames, String[] opponentNames) {
         if (view != null) {
             this.view.setNames(partyNames, opponentNames);
         }
@@ -77,19 +56,7 @@ public class BattlePresenter {
      * @return whether the given name corresponds to a fainted character.
      */
     public boolean givenCharacterFainted(String name) {
-        for (BattleCharacter c : encounter.getPlayerCharacters()) {
-            if (c.getName().equals(name)) {
-                return false;
-            }
-        }
-
-        for (BattleCharacter c : encounter.getOpponents()) {
-            if (c.getName().equals(name)) {
-                return false;
-            }
-        }
-
-        return true;
+        return interactor.isCharacterFainted(name);
     }
 
     /**
@@ -98,7 +65,7 @@ public class BattlePresenter {
      * @return the desired BattleCharacter's current health.
      */
     public int givenCharacterHealth(String name) {
-        return this.encounter.getCharacter(name).getHp();
+        return interactor.getCharacterHealth(name);
     }
 
     /**
@@ -107,7 +74,7 @@ public class BattlePresenter {
      * @return the desired BattleCharacter's current damage.
      */
     public int givenCharacterDamage(String name) {
-        return this.encounter.getCharacter(name).getDmg();
+        return interactor.getCharacterDamage(name);
     }
 
     /**
@@ -117,9 +84,7 @@ public class BattlePresenter {
      * @return the desired BattleCharacter's move stats.
      */
     public int[] givenCharacterMoveStats(String name) {
-        BattleCharacter caster = this.encounter.getCharacter(name);
-        return new int[] {caster.getMoveOne().getHealthChange(), caster.getMoveOne().getDamageChange(),
-            caster.getMoveTwo().getHealthChange(), caster.getMoveTwo().getDamageChange()};
+        return interactor.getCharacterMoveStats(name);
     }
 
     /**
@@ -129,8 +94,7 @@ public class BattlePresenter {
      * @return the desired BattleCharacter's move names.
      */
     public String[] givenCharacterMoveNames(String name) {
-        BattleCharacter caster = this.encounter.getCharacter(name);
-        return new String[] {caster.getMoveOne().getName(), caster.getMoveTwo().getName()};
+        return interactor.getCharacterMoveNames(name);
     }
 
     /**
@@ -141,48 +105,7 @@ public class BattlePresenter {
      * @return a String of the name of the moving character
      */
     public String roundStart() {
-
-        int status = this.encounter.getBattleStatus();
-        if (status == -1) { //Player lost last round
-            //Something with ReplayGenerator
-            return null;
-        } else if (status == 1) { //Player won last round
-            this.endBattle();
-            return null;
-        } else { //Battle is ongoing
-            BattleCharacter moving = this.encounter.getMovingCharacter();
-
-            if (moving.isOpponent()) { //Opponent character is moving
-                Random rand = new Random();
-                int moveNumber = rand.nextInt(2);
-
-                Move chosenMove;
-                if (moveNumber == 0) {
-                    chosenMove = moving.getMoveOne();
-                } else {
-                    chosenMove = moving.getMoveTwo();
-                }
-
-                if (!chosenMove.isFriendly()) {
-                    ArrayList<BattleCharacter> players = this.encounter.getPlayerCharacters();
-
-                    //Choose a random player character to attack
-                    int target = rand.nextInt(players.size());
-
-                    this._useMove(chosenMove, moving, players.get(target));
-                } else {
-                    ArrayList<BattleCharacter> opponents = this.encounter.getOpponents();
-
-                    //Choose a random opponent to use the friendly move on
-                    int target = rand.nextInt(opponents.size());
-
-                    this._useMove(chosenMove, moving, opponents.get(target));
-                }
-            }
-
-            //No matter what, the moving BattleCharacter's name is returned
-            return moving.getName();
-        }
+        return interactor.roundStart();
     }
 
     /**
@@ -195,34 +118,7 @@ public class BattlePresenter {
      * @return ArrayList of Strings representing the names of every possible target BattleCharacter.
      */
     public ArrayList<String> retrieveTargets(int moveNum, String casterName) {
-        //Retrieve the moving BattleCharacter
-        BattleCharacter caster = this.encounter.getCharacter(casterName);
-
-        //Retrieve the Move being used
-        Move m;
-        if (moveNum == 1) {
-            m = caster.getMoveOne();
-        } else { //moveNum == 2
-            m = caster.getMoveTwo();
-        }
-
-        ArrayList<String> targets = new ArrayList<>();
-
-        ArrayList<BattleCharacter> targetCharacters;
-
-        //Retrieve possible target BattleCharacters
-        if (m.isFriendly()) { //caster is a player character, so targets are player characters
-            targetCharacters = this.encounter.getPlayerCharacters();
-        } else { //caster is a player character, so targets are opponent characters
-            targetCharacters = this.encounter.getOpponents();
-        }
-
-        //Add character names to ArrayList
-        for (BattleCharacter c : targetCharacters) {
-            targets.add(c.getName());
-        }
-
-        return targets;
+        return interactor.retrieveTargets(moveNum, casterName);
     }
 
     /**
@@ -232,55 +128,24 @@ public class BattlePresenter {
      * @param targetName String representing the name of the target BattleCharacter.
      */
     public void executeTurn(int moveNum, String casterName, String targetName) {
-        BattleCharacter caster = this.encounter.getCharacter(casterName);
-        BattleCharacter target = this.encounter.getCharacter(targetName);
-        if (moveNum == 1) {
-            this._useMove(caster.getMoveOne(), caster, target);
-        } else { //moveNum == 2
-            this._useMove(caster.getMoveTwo(), caster, target);
-        }
-    }
-
-    public void endBattle() {
-        //unimplemented, includes _removeActiveBattle and _addReward()
+        interactor.executeTurn(moveNum, casterName, targetName);
     }
 
     /**
-     * Applies the given move used by given caster onto the given target.
-     * If the target faints and is a player character, remove it from the list of active players in state,
-     * and add it to the list of fainted characters.
-     * Calls on the view to update the impacted character.
+     * UNIMPLEMENTED
      */
-    private void _useMove(Move m, BattleCharacter caster, BattleCharacter target) {
-        target.modifyHealth(m.getHealthChange());
+    public void endBattle() {
+        interactor.endBattle();
+    }
 
-        //If the move is to be used on a character on the other team, and the move is meant to deal health damage,
-        // then increase the health damage applied by the move based on the character's damage stat.
-        if (!m.isFriendly() && m.getHealthChange() < 0) {
-            target.modifyHealth(-caster.getDmg());
-        }
-
-        if (target.getHp() == 0) {
-            this.encounter.removeChar(target);
-            if (!target.isOpponent()) {
-                this.state.getParty().remove(target);
-                this.state.getFainted().add(target);
-            }
-        }
-
-        target.modifyDamage(m.getDamageChange());
-
+    /**
+     * Call on the BattleMenuInterface to update the displayed information for the given character.
+     * @param targetName the name String of the character whose information displayed on the view needs to be updated.
+     */
+    @Override
+    public void updateViewCharacter(String targetName) {
         if (view != null) {
-            this.view.updateCharacter(target.getName());
+            this.view.updateCharacter(targetName);
         }
     }
-
-    private void _addReward() {
-        //unimplemented
-    }
-
-    private void _removeActiveBattle() {
-        //unimplemented
-    }
-
 }
